feat: add incoming deposits to wallet screen
